Index

Conservation

ExoFlow.conservations_equations — Function

conservations_equations(params::Dict{Symbol,Any})

It provides the Navier Stokes equations: continuity and momentum. The equations are different if the problem is linearized or not. Conitnuity: $\nabla\cdot u = 0$

Momentum: $\dfrac{\partial u}{\partial t} + u \cdot \nabla(u) + \nabla (p) + \nu \Delta u= 0$

Momentum linearized: $\tilde{u} \cdot \nabla(u) + \nabla (p) + \nu \Delta u = 0$ where $\tilde{u}$ is the approximation of the convective velocity.

ExoFlow.derivative_conservations_equations — Function

derivative_conservations_equations(params::Dict{Symbol,Any})

It provides the spatial derivative of the Navier Stokes equations: continuity and momentum. The equations are different if the problem is linearized or not. Conitnuity: $\nabla\cdot du = 0$

Momentum: $du \cdot \nabla(u) + u \cdot \nabla(du) + \nabla (dp) + \nu \Delta du= 0$

Momentum linearized: $\tilde{u} \cdot \nabla(u) + \nabla (p) + \nu \Delta u = 0$

Linear Utilities

ExoFlow.create_ũ_vector — Function

It instantiate an AbstactVector of 4 elements. Each element is an AbstractVector where the values of the velocity in each node is stored. The first element refers to the time step n, the second to the time step n-1, the third n-2 and the last n-3.

createũvector(zfields::SequentialData)

ExoFlow.update_ũ — Function

It updates the convective velocity exitmation $\tilde{u}$ for the approximation of the non linear term: $\tilde{u} \cdot \nabla(u)$

ExoFlow.update_ũ_vector! — Function

It updates the vector which stores the values of velocity at previous time steps.

ExoFlow.update_linear! — Function

update_linear!(params::Dict{Symbol,Any})

Wrapper function for updating the ũ_vector containing the speed values at previous time steps and ũ which is the approximation of ũ for the non linear terms, ũ⋅(∇u)

SUPG

SUPG

PETSc

ExoFlow.petsc_options — Function

petsc_options(prec::Symbol)

The function has some solver options for PETSc solver. The user can set the argument of the function:

- `:snesgamg` for using a Geometric Algebraic MultiGrid preconditioner and the Newton method for solving a non linear problem
- `:kspgamg` for using a Geometric Algebraic MultiGrid preconditioner and the gmres method for solving a linear problem
- `:sneslu` for using an Additive Shwarz Method as preconditioner and a LU factorization as sub-preconditioner and the Newton method for solving a non linear problem
- `:ksplu` for using an Additive Shwarz Method as preconditioner and a LU factorization as sub-preconditioner and the gmres method for solving a linear problem

The gamg preconditioner is not compatible with the SequentialBackend()

Common procedures