Index

Conservation

ExoFlow.conservations_equations — Function

conservations_equations(params::Dict{Symbol,Any})

It provides the Navier Stokes equations: continuity and momentum. The equations are different if the problem is linearized or not. Conitnuity: $\nabla\cdot u = 0$

Momentum: $\dfrac{\partial u}{\partial t} + u \cdot \nabla(u) + \nabla (p) + \nu \Delta u= 0$

Momentum linearized: $\tilde{u} \cdot \nabla(u) + \nabla (p) + \nu \Delta u = 0$ where $\tilde{u}$ is the approximation of the convective velocity.

ExoFlow.derivative_conservations_equations — Function

derivative_conservations_equations(params::Dict{Symbol,Any})

It provides the spatial derivative of the Navier Stokes equations: continuity and momentum. The equations are different if the problem is linearized or not. Conitnuity: $\nabla\cdot du = 0$

Momentum: $du \cdot \nabla(u) + u \cdot \nabla(du) + \nabla (dp) + \nu \Delta du= 0$

Momentum linearized: $\tilde{u} \cdot \nabla(u) + \nabla (p) + \nu \Delta u = 0$

Linear Utilities

ExoFlow.create_ũ_vector — Function

It instantiate an AbstactVector of 4 elements. Each element is an AbstractVector where the values of the velocity in each node is stored. The first element refers to the time step n, the second to the time step n-1, the third n-2 and the last n-3.

createũvector(zfields::SequentialData)

ExoFlow.update_ũ — Function

It updates the convective velocity exitmation $\tilde{u}$ for the approximation of the non linear term: $\tilde{u} \cdot \nabla(u)$

ExoFlow.update_ũ_vector! — Function

It updates the vector which stores the values of velocity at previous time steps.

ExoFlow.update_linear! — Function

update_linear!(params::Dict{Symbol,Any})

Wrapper function for updating the ũ_vector containing the speed values at previous time steps and ũ which is the approximation of ũ for the non linear terms, ũ⋅(∇u)

SUPG

ExoFlow.h_param — Function

It computes the dimension of each cell of the computational domain. It uses the approximation: $h = M^{1/D}$ where h is the cell dimension, M the cell measure (surface for 2D, volume for 3D) and D is the Dimension (2 or 3)

ExoFlow.τ — Function

Stabilization parameter for momentum equation for the SUPG.

ExoFlow.τb — Function

Stabilization parameter for continuity equation for the SUPG.

ExoFlow.SUPG — Function

SUPG non linear variational forumulation. It calles the conservations_equations and derivative_conservations_equations. It provides the equations set, the jacobian and the time jacobian.

ExoFlow.SUPG_lin — Function

SUPG linear variational forumulation. It calles the conservations_equations and derivative_conservations_equations. The terms are divided between mass matrix, stifness matrix and right hand side.

VMS

ExoFlow.G_params — Function

It provides the set of stabilization parameters for the VMS: G, GG, gg. GG comes from the inverse of the jacobian of the map between reference and physical domain. It is exactly the cell dimension obtained by h_param for orthogonal carthesian grids.

ExoFlow.τm — Function

Stabilization parameter for momentum equation for the VMS.

ExoFlow.τc — Function

Stabilization parameter for contitnuity equation for the VMS.

ExoFlow.VMS — Function

VMS non linear variational forumulation. It calles the conservations_equations and derivative_conservations_equations. It provides the equations set, the jacobian and the time jacobian.

ExoFlow.VMS_lin — Function

VMS linear variational forumulation. It calles the conservations_equations and derivative_conservations_equations. The terms are divided between mass matrix, stifness matrix and right hand side.

PETSc

ExoFlow.petsc_options — Function

petsc_options(prec::Symbol)

The function has some solver options for PETSc solver. The user can set the argument of the function:

- `:snesgamg` for using a Geometric Algebraic MultiGrid preconditioner and the Newton method for solving a non linear problem
- `:kspgamg` for using a Geometric Algebraic MultiGrid preconditioner and the gmres method for solving a linear problem
- `:sneslu` for using an Additive Shwarz Method as preconditioner and a LU factorization as sub-preconditioner and the Newton method for solving a non linear problem
- `:ksplu` for using an Additive Shwarz Method as preconditioner and a LU factorization as sub-preconditioner and the gmres method for solving a linear problem

The gamg preconditioner is not compatible with the SequentialBackend()

Common procedures

ExoFlow.hf_gen! — Function

Add the body force. The user in the initial parameters specify the component in the x direction. Then the body force function is created 2D or 3D accordingly to the physics of the case.

ExoFlow.add_centre_tag! — Function

It creates the 'center' tag at the tag_coordinate (Point); if mesh extremely fine the tolrances have to be smaller (unlikely)

ExoFlow.printmodel — Function

It prints the mesh of the model

ExoFlow.creation_fe_spaces — Function

It creates the finite elements spaces accordingly to the previously generated dirichelet tags

ExoFlow.creation_op — Function

It creates the problem, here are implemented the VMS or SUPG system of equations

ExoFlow.creation_initial_conditions — Function

It creates the initial conditions. The default initial value for the pressure is 0.0 Pa. If the restart parameter is set to true, the initial conditons are set accordingly to the provided file

ExoFlow.create_system_solver — Function

It creates the system solver which can be linear/non linear, julia/petsc using GAMG (only in petsc) or not.

ExoFlow.creation_ode_parameters — Function

It creates parameters that are used for the ODE solution. In specific, if different $\theta$ are specified for velocity and pressure, are created vectors which relate each degree of freedom to the desired $\theta$ .

ExoFlow.creation_ode_solver — Function

It creates the ODE solver using the parameters defined by the user

ExoFlow.compute_solution_ — Function

It computes the solution, different for benchmark case or for getting the results.

ExoFlow.compute_solution — Function

It computes the solution, in the serial or parallel case.

ExoFlow.iterate_solution — Function

It iterates the solution, here is where the solution is actually computed. At the end of each step the forces values are extracted if required

ExoFlow.compute_solution_benchmark — Function

For benchmarking. It removes the compilation time for the first iteration and it does not print any results.

ExoFlow.solve_case — Function

Wrapper function that solve the case. Here the Non Linear Solver - nls is defined, in case the default julia sovler is adopted or PETSc

ExoFlow.initialize_parameters — Function

initialize_parameters(params::Dict{Symbol, Any})

It makes some checks on the validity of the input provided by the user and add some additional information.

Restart

ExoFlow.find_idx — Function

find_idx(p::VectorValue{2, Float64}, params; atol=1e-6)

For a given point p it search on the restart_file provided by the user the closer node and provides the index of such node. It takes into account also the channel periodic case, changing the coordiantes of the periodic directions.

find_idx(p::VectorValue{3, Float64}, params; atol = 1e-4)

For a given point p it search on the restart_file provided by the user the closer node and provides the index of such node. It takes into account also the channel periodic case, changing the coordiantes of the periodic directions.

ExoFlow.uh_restart — Function

uh_restart(p, params::Dict{Symbol, Any})

For a given point p it calles find_idx which provide the line of the csv file corresponding to that point. Then, it calles uh which provide the VectorValue of the velocity at that point.

ExoFlow.ph_restart — Function

ph_restart(p, params::Dict{Symbol, Any})

For a given point p it calles find_idx which provide the line of the csv file corresponding to that point. Then, it calles ph which provide the scalar pressure value at that point.

ExoFlow.restart_uh_field — Function

restart_uh_field(params::Dict{Symbol, Any})

It provides a suitable function which gives for each point the specified velocity in restart_file. It is used as initial condition for restarting a simulation at a specific time step.

ExoFlow.restart_ph_field — Function

restart_ph_field(params::Dict{Symbol, Any})

It provides a suitable function which gives for each point the specified pressure in restart_file. It is used as initial condition for restarting a simulation at a specific time step.

Forces

ExoFlow.forces_domain — Function

forces_domain(model, force_tags, degree)

For a given force_tags in the model it provides the triangulation, the measure and the normal vector.

ExoFlow.compute_forces — Function

compute_forces(p, u, n_Γ, dΓ, ρ, ν, tn)

It computes the pressure and friction forces.

ExoFlow.write_forces — Function

write_forces(params::Dict{Symbol, Any}, uh_tn, ph_tn, tn::Float64)

It writes on a csv file the forces over the body (airfoil or cylinder). It writes 5 columns: time step, pressure force in x direction, pressure force in y direction, friction force in x direction, friction force in y direction.